application/consumer/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9
pika==1.3.2
pytz==2023.3 